+++ cypress/integration/allEnv/configuration/channels.js
@@ -2,14 +2,13 @@
 import faker from "faker";
 
 import { createChannel } from "../../../apiRequests/Channels";
-import { LEFT_MENU_SELECTORS } from "../../../elements/account/left-menu/left-menu-selectors";
+import { ONE_PERMISSION_USERS } from "../../../Data/users";
 import { PRODUCTS_LIST } from "../../../elements/catalog/products/products-list";
 import { ADD_CHANNEL_FORM_SELECTORS } from "../../../elements/channels/add-channel-form-selectors";
 import { AVAILABLE_CHANNELS_FORM } from "../../../elements/channels/available-channels-form";
 import { CHANNEL_FORM_SELECTORS } from "../../../elements/channels/channel-form-selectors";
 import { CHANNELS_SELECTORS } from "../../../elements/channels/channels-selectors";
 import { SELECT_CHANNELS_TO_ASSIGN } from "../../../elements/channels/select-channels-to-assign";
-import { CONFIGURATION_SELECTORS } from "../../../elements/configuration/configuration-selectors";
 import { HEADER_SELECTORS } from "../../../elements/header/header-selectors";
 import { DRAFT_ORDER_SELECTORS } from "../../../elements/orders/draft-order-selectors";
 import { ORDERS_SELECTORS } from "../../../elements/orders/orders-selectors";
@@ -29,17 +28,10 @@ describe("Channels", () => {
   });
 
   beforeEach(() => {
-    cy.clearSessionData().loginUserViaRequest();
-  });
-
-  it("should navigate to channels page", () => {
-    cy.visit(urlList.homePage)
-      .get(LEFT_MENU_SELECTORS.configuration)
-      .click()
-      .get(CONFIGURATION_SELECTORS.channels)
-      .click()
-      .location("pathname")
-      .should("contain", "channels");
+    cy.clearSessionData().loginUserViaRequest(
+      "auth",
+      ONE_PERMISSION_USERS.channel
+    );
   });
 
   it("should create new channel", () => {
@@ -65,6 +57,7 @@ describe("Channels", () => {
       .click();
 
     // new channel should be visible at product availability form
+    cy.clearSessionData().loginUserViaRequest();
     cy.addAliasToGraphRequest("InitialProductFilterAttributes");
     cy.visit(urlList.products);
     cy.wait("@InitialProductFilterAttributes");
@@ -135,6 +128,7 @@ describe("Channels", () => {
       slug: randomChannel,
       currencyCode: currency
     });
+    cy.clearSessionData().loginUserViaRequest();
     cy.visit(urlList.orders)
       .get(ORDERS_SELECTORS.createOrder)
       .click()