+++ cypress/integration/allEnv/products/updatingProducts.js
@@ -3,6 +3,7 @@ import faker from "faker";
 import { createCategory } from "../../../apiRequests/Category";
 import { createCollection } from "../../../apiRequests/Collections";
 import { getProductDetails } from "../../../apiRequests/storeFront/ProductDetails";
+import { ONE_PERMISSION_USERS } from "../../../Data/users";
 import { PRODUCT_DETAILS } from "../../../elements/catalog/products/product-details";
 import { BUTTON_SELECTORS } from "../../../elements/shared/button-selectors";
 import { metadataForms } from "../../../steps/catalog/metadataSteps";
@@ -56,6 +57,7 @@ describe("Update products", () => {
         product = productResp;
       });
   });
+
   it("Should update product", () => {
     const updatedName = `${startsWith}${faker.random.number()}`;
     let updatedCategory;
@@ -95,8 +97,11 @@ describe("Update products", () => {
             collection: updatedCollection.name
           }
         };
-        cy.visit(productDetailsUrl(product.id));
-        cy.get(PRODUCT_DETAILS.collectionRemoveButtons).click();
+        cy.clearSessionData()
+          .loginUserViaRequest("auth", ONE_PERMISSION_USERS.product)
+          .visit(productDetailsUrl(product.id))
+          .get(PRODUCT_DETAILS.collectionRemoveButtons)
+          .click();
         fillUpCommonFieldsForAllProductTypes(productData, false);
         cy.addAliasToGraphRequest("UpdatePrivateMetadata");
         cy.addAliasToGraphRequest("UpdateMetadata");
@@ -122,15 +127,18 @@ describe("Update products", () => {
           });
       });
   });
+
   it("should delete product", () => {
-    cy.visit(productDetailsUrl(product.id));
-    cy.addAliasToGraphRequest("ProductDelete");
-    cy.get(BUTTON_SELECTORS.deleteButton)
+    cy.clearSessionData()
+      .loginUserViaRequest("auth", ONE_PERMISSION_USERS.product)
+      .visit(productDetailsUrl(product.id))
+      .addAliasToGraphRequest("ProductDelete")
+      .get(BUTTON_SELECTORS.deleteButton)
       .click()
       .get(BUTTON_SELECTORS.submit)
-      .click();
-    cy.wait("@ProductDelete");
-    cy.loginUserViaRequest("token")
+      .click()
+      .wait("@ProductDelete")
+      .loginUserViaRequest("token")
       .then(() => {
         getProductDetails(product.id, defaultChannel.slug).its("body.data");
       })