+++ src/components/SingleAutocompleteSelectField/SingleAutocompleteSelectField.tsx
@@ -46,7 +46,7 @@ export interface SingleAutocompleteSelectFieldProps
   InputProps?: InputProps;
   fetchChoices?: (value: string) => void;
   onChange: (event: React.ChangeEvent<any>) => void;
-  onFocus?: () => void;
+  fetchOnFocus?: boolean;
   FormHelperTextProps?: ExtendedFormHelperTextProps;
   nakedInput?: boolean;
 }
@@ -76,7 +76,7 @@ const SingleAutocompleteSelectFieldComponent: React.FC<SingleAutocompleteSelectF
     fetchChoices,
     onChange,
     onFetchMore,
-    onFocus,
+    fetchOnFocus,
     FormHelperTextProps,
     nakedInput = false,
     ...rest
@@ -170,8 +170,8 @@ const SingleAutocompleteSelectFieldComponent: React.FC<SingleAutocompleteSelectF
               onBlur: handleBlur,
               onClick: toggleMenu,
               onFocus: () => {
-                if (onFocus) {
-                  onFocus();
+                if (fetchOnFocus) {
+                  fetchChoices(inputValue);
                 }
               }
             };