+++ src/fragments/types/CategoryTranslationFragment.ts
@@ -17,7 +17,7 @@ export interface CategoryTranslationFragment_translation {
   id: string;
   description: any | null;
   language: CategoryTranslationFragment_translation_language;
-  name: string;
+  name: string | null;
   seoDescription: string | null;
   seoTitle: string | null;
 }