+++ src/products/views/ProductList/fixtures.ts
@@ -5,14 +5,12 @@ import { fetchMoreProps, searchPageProps } from "@saleor/fixtures";
 import { ProductListFilterOpts } from "@saleor/products/components/ProductListPage";
 import { productTypes } from "@saleor/productTypes/fixtures";
 import { StockAvailability } from "@saleor/types/globalTypes";
+import { mapEdgesToItems, mapSlugNodeToChoice } from "@saleor/utils/maps";
 
 export const productListFilterOpts: ProductListFilterOpts = {
   attributes: attributes.map(attr => ({
+    id: attr.id,
     active: false,
-    choices: attr.choices.edges.map(val => ({
-      label: val.node.name,
-      value: val.node.slug
-    })),
     name: attr.name,
     slug: attr.slug,
     value: [
@@ -20,6 +18,14 @@ export const productListFilterOpts: ProductListFilterOpts = {
       attr.choices.edges.length > 2 && attr.choices.edges[2].node.slug
     ]
   })),
+  attributeChoices: {
+    ...fetchMoreProps,
+    ...searchPageProps,
+    active: false,
+    value: null,
+    choices: mapSlugNodeToChoice(mapEdgesToItems(attributes[0].choices)),
+    displayValues: mapSlugNodeToChoice(mapEdgesToItems(attributes[0].choices))
+  },
   categories: {
     ...fetchMoreProps,
     ...searchPageProps,