+++ src/storybook/stories/products/ProductVariantCreatePage.tsx
@@ -40,7 +40,6 @@ storiesOf("Views / Products / Create product variant", module)
       fetchAttributeValues={() => undefined}
       onAssignReferencesClick={() => undefined}
       onCloseDialog={() => undefined}
-      onAttributeFocus={() => undefined}
     />
   ))
   .add("with errors", () => (
@@ -83,7 +82,6 @@ storiesOf("Views / Products / Create product variant", module)
       fetchAttributeValues={() => undefined}
       onAssignReferencesClick={() => undefined}
       onCloseDialog={() => undefined}
-      onAttributeFocus={() => undefined}
     />
   ))
   .add("when loading data", () => (
@@ -107,7 +105,6 @@ storiesOf("Views / Products / Create product variant", module)
       fetchAttributeValues={() => undefined}
       onAssignReferencesClick={() => undefined}
       onCloseDialog={() => undefined}
-      onAttributeFocus={() => undefined}
     />
   ))
   .add("add first variant", () => (
@@ -134,7 +131,6 @@ storiesOf("Views / Products / Create product variant", module)
       fetchAttributeValues={() => undefined}
       onAssignReferencesClick={() => undefined}
       onCloseDialog={() => undefined}
-      onAttributeFocus={() => undefined}
     />
   ))
   .add("no warehouses", () => (
@@ -158,6 +154,5 @@ storiesOf("Views / Products / Create product variant", module)
       fetchAttributeValues={() => undefined}
       onAssignReferencesClick={() => undefined}
       onCloseDialog={() => undefined}
-      onAttributeFocus={() => undefined}
     />
   ));