+++ src/storybook/stories/products/ProductVariantPage.tsx
@@ -39,7 +39,6 @@ storiesOf("Views / Products / Product variant details", module)
       fetchAttributeValues={() => undefined}
       onAssignReferencesClick={() => undefined}
       onCloseDialog={() => undefined}
-      onAttributeFocus={() => undefined}
     />
   ))
   .add("when loading data", () => (
@@ -68,7 +67,6 @@ storiesOf("Views / Products / Product variant details", module)
       fetchAttributeValues={() => undefined}
       onAssignReferencesClick={() => undefined}
       onCloseDialog={() => undefined}
-      onAttributeFocus={() => undefined}
     />
   ))
   .add("no warehouses", () => (
@@ -96,7 +94,6 @@ storiesOf("Views / Products / Product variant details", module)
       fetchAttributeValues={() => undefined}
       onAssignReferencesClick={() => undefined}
       onCloseDialog={() => undefined}
-      onAttributeFocus={() => undefined}
     />
   ))
   .add("attribute errors", () => (
@@ -152,6 +149,5 @@ storiesOf("Views / Products / Product variant details", module)
       fetchAttributeValues={() => undefined}
       onAssignReferencesClick={() => undefined}
       onCloseDialog={() => undefined}
-      onAttributeFocus={() => undefined}
     />
   ));